+++ test/beast/core/bind_handler.cpp
@@ -11,7 +11,7 @@
 #include <boost/beast/core/bind_handler.hpp>
 
 #include <boost/beast/core/detail/type_traits.hpp>
-#include <boost/beast/test/stream.hpp>
+#include <boost/beast/experimental/test/stream.hpp>
 #include <boost/beast/unit_test/suite.hpp>
 #include <boost/asio/io_service.hpp>
 #include <boost/asio/strand.hpp>