+++ test/beast/core/buffered_read_stream.cpp
@@ -11,7 +11,7 @@
 #include <boost/beast/core/buffered_read_stream.hpp>
 
 #include <boost/beast/core/multi_buffer.hpp>
-#include <boost/beast/test/stream.hpp>
+#include <boost/beast/experimental/test/stream.hpp>
 #include <boost/beast/test/yield_to.hpp>
 #include <boost/beast/unit_test/suite.hpp>
 #include <boost/asio/buffer.hpp>