+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 170:
 * Add ssl_stream to experimental
 * Add test::error to experimental
 * Add test::fail_count to experimental
+* Add test::stream to experimental
 
 --------------------------------------------------------------------------------